For ingredients I added the basic route actions and views. I thought I would add similar functionality to the recipe ingredients as to the userrecipes, but then I realized that saved recipes were just duplicate links. I can't actually change the recipes as they are now or else I would be changing all saved copies of that recipe of all users.
I will probably have to fundamentally alter the database structure and table relationships of the underlying models and create a new model type or join the two join tables.

In the meantime I am attempting to hack around the issue by writing methods for the saved recipes(UserRecipes) that mimics the methods of regular recipe objects and then I am going to attempt to simply add new ingredients to the ingredients on there. I am quite sure this will be a problem though.

I rewrote the user association with userrecipes to be saved recipes.
I changed the views accordingly.